+++ src/main/java/racingcar/model/carnameinput/CarNameSplit.java
@@ -1,4 +1,4 @@
-package racingcar.model.carname;
+package racingcar.model.carnameinput;
 
 public class CarNameSplit {
     private static final String DELIMITER = ",";
