+++ src/test/java/racingcar/model/carnameinput/CarNameSplitTest.java
@@ -1,4 +1,4 @@
-package racingcar.model.carname;
+package racingcar.model.carnameinput;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
